Created first document
I created the first proposal document. If any of you (Haravikk, Nate,
Jeff) wish to be added as authors please do so, I didn’t want to add
anyone without consent. We still need to add the code in Detailed
Design, as we keep improving it, and then we can pull this to the main
swift rep.